Key value pairs are not sorted on key values.
The underlying mechanism used to display a map of values, sorted the values in the map, as well as collapsed keys.  The new format now allows for multiple keys to be displayed in order as they presented in the backing file or url.

I did notice that the list values that are entered inline via the GUI are sorted.  This needs to be investigated further.

[#5 state:resolved responsible:'Noah Campbell']