Add alternate, safe quoting to CLIUtils (optional, off-by-default)
This is the (very) beginning of a fix to #298.

Unfortunately, doing this in a way that doesn't compromise backwards
compatibility is out-of-scope, and will probably call for a multi-phase
implementation -- for instance, adding a per-job flag to specify
compatible or safe quoting, providing warnings and (eventually)
deprecation notices to users with jobs using unsafe quoting, and (again,
eventually) flipping the default.