Added support for nested groups in JettyCachingLdapLoginModule. It will download all groups and map out which groups are members of other groups, and then recursively build the list of all groups a user is member of.
The structure with which groups are members of other groups is cached
for the same duration as the caching of UserInfo. The same structure is
shared for all users, so it will only be built once for all users.

The feature is enabled/disabled with the new option nestedGroups in the
jaas config.