Add threadcount as a job context variable
This commit adds threadcount as a job context variable, so scripts and
commands called by rundeck can use the threadcount of a job. For
example, if you are running a job that involves taking some nodes in a
cluster offline, the threadcount vs. the total number of nodes in the
cluster could be used by called scripts to check that enough capacity
will be left.