Fix scrolling in log output
Fixes #4047.

In 2.x the log output window scrolled to the bottom on new additions to
the table. This scrolls the window, but in the new 3.x UI the main
scrolling element is actually a div named #main-panel.

Instead of scrolling the window's scrolling area (which is fixed, so
that the navigation on the left doesn't move), we instead scroll
the main-panel element.