ENH: Support Arbitrary User IDs on OpenShift
PROBLEM:
OpenShift Container Platform runs containers using an arbitrarily assigned user ID, this means the rundeck process will be started with an UID not equal to the rundeck user defined in the passwd. This leads to problems, for example using ssh.

SOLUTION:
simple add a a new rundeck passwd entry with the correct UID during the start of the container. 
Of course the /etc/passwd needs write permissions for the group (-> adding RUN chmod g=u /etc/passwd to the Dockerfile as well)

Morge background information:
https://docs.openshift.com/container-platform/3.3/creating_images/guidelines.html#openshift-container-platform-specific-guidelines