Allow several tokens per user for API authentication
This commit will ease the rotation of tokens used to authenticate to
the Rundeck API.

Currently the token used by the clients of the Rundeck API mut be in
sync with the one configured in the static file of Rundeck, which
creates a service disruption.  With this commit, the clients will be
able to switch to a new token without disruption.

A later commit in rundeck/docs will mention this feature in the
documentation.