fix: Error page shows wrong message
If an exception is thrown, it seems the handling mechanism invokes the ProjectSelectInterceptor a 2nd time, and also causes the "project" parameter to be doubled, resulting in `params.project` being a String[] type with two string values, each equal to the project name.  This causes the ProjectSelectInterceptor to incorrectly evaluate the project name as toString value of the String[], and show a "Project Name is invalid" message instead of the true error.  This fix resets the parameter value to a single String if params.project is a String[]