Add bootstrap storage tree rewrite feature
this feature will automatically re-store any key storage resource that is not encrypted,
when the jasypt-encryption converter plugin is in use. If the plugin is not configured,
this feature will not do anything.  If it is in use, and the enabled flag is set,
each resource that is not already encrypted, will be restored to encrypt it.

configuration:

enabled (default false):

    rundeck.feature.storageRewrite.enabled=true

base path (default "keys"):

    rundeck.storage.rewrite.basePath=keys