Fix flaky test for avg duration notification
I believe this was flaky because it relied on a timing issue
the ExecutionJob calls thread.join(1000), however the
thread object was a spock Mock, so the default behavior is return
immediately.  This was causing the average duration exceeded check
to only sometimes fail, because the average duration would be set
to 1ms, and the notification method not be called.

the fix replaces the Mock with a real thread that waits for a
countdownlatch, causing the thread.join(1000) to actually wait for
1000ms, guaranteeing the average duration will exceed, and the
notification method will be called. when the averge notification method
is called, the countdownlatch is set to 0 allowing the real thread
to finish