RCLOUD-1413 Remove optional template invocation from login page
The code change will remove a template invocation in login.gsp, the
login page. As far as we know it is only used for a rundeckpro
functionality that is being removed.

There is a related commit for this same issue on rundeckpro.

See original changes that introduced this on the UI on RCLOUD-561 and
RCLOUD-596.