fix: ignoreRoles option should work even if binding login context is used
problem:
the JettyCombinedLdapLoginModule cannot be used with the JettyRolePropertyFileLoginModule to supply roles if a binding context is used, because the ignoreRoles attribute is not honored.

This solution allows JettyCombinedLdapLoginModule to honor the ignoreRoles attribute in both cases, by overriding the getUserRolesByDn instead of the getUserRoles method.  The former is called by the binding context flow, whereas the latter was only called when in a non-binding context. the solution still works with non-binding context because the getUserRoles method subsequently calls getUserRolesByDn as well.