Test: Disable all HTTP method filtering to isolate ServletApiRequest NPE
Temporarily disabled BanHttpMethodCustomizer (both @Configuration and filter)
to determine if HTTP method filtering is causing the ServletApiRequest.getRequest()
returning null issue.

If this resolves the NPE, we know HTTP method filtering is the culprit and need
a different approach for Jetty 12.

If NPE persists, the issue is elsewhere in the Jetty/Spring configuration.